Fix Newton-Rapshon iteration
In the previous version, some equations of the linear system were divided by lams, while other were not. The easiest to see why this is wrong, is by considering the solution of the NON-LINEAR system of equations F(x)=0. Since this system is non-linear, you cannot divide some rows by lams, while not doing this to other rows. To be consistent, the full resulting equations from the symbolic equations is now implemented